+++ MovieQuiz/Services/StatisticServiceProtocol.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-protocol StatisticService {
+protocol StatisticServiceProtocol {
     func store(correct count: Int, total amount: Int)
     var totalAccuracy: Double { get }
     var gamesCount: Int { get }
